+++ src/main/java/entities/RecurrentEvent.java
@@ -1,28 +1,66 @@
+/*
+This is the RecurrentEvent Entity;
+
+ */
+
 package entities;
+
+import java.sql.Time;
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 
-// Entity layer
 
-public class RecurrentEvent extends Event {
+public class RecurrentEvent implements Event {
 
-    private boolean isBiweekly;
+    private String name;
+    private LocalDateTime startTime;
+    private LocalDateTime endTime;
+    private boolean isCommute;
+    private String location;
+    private Event commute;
 
-    /**
-     * Constructs a RecurringEvent with a name, startTime, endTime, commuteTime, and location.
-     *
-     * @param name the name of the event to be created
-     * @param startTime the start time of the event to be created
-     * @param endTime the end time of the event to be created
-     * @param commuteTime the time it takes to get to this event
-     * @param location the location of the event to be created
-     */
-    public RecurrentEvent(String name, LocalDateTime startTime, LocalDateTime endTime, Boolean isCommute, Event commuteTime, String location) {
-        super(name, startTime, endTime, isCommute, commuteTime, location);
-        this.isBiweekly = false;
+
+    public RecurrentEvent(String name, LocalDateTime startTime, LocalDateTime endTime, boolean isCommute, Event commute, String location){
+
+        this.name = name;
+        this.startTime = startTime;
+        this.endTime = endTime;
+        this.commute = commute;
+        this.location = location;
+        this.isCommute = isCommute;
     }
 
+    @Override
+    public String getName() {
+        return this.name;
+    }
+
+    @Override
+    public Event getCommute() {
+        return this.commute;
+    }
+    @Override
+    public LocalDateTime getStartTime() {
+        return startTime;
+    }
+    @Override
+    public LocalDateTime getEndTime() {
+        return endTime;
+    }
+    @Override
+    public String getLocation() {
+        return location;
+    }
+    @Override
+    public boolean isCommute(){
+        return this.isCommute;
+    }
+
+    @Override
+    public boolean isValid() {
+        if (endTime.isBefore(startTime)) {
+            return false;
+        }
+        return true;
+    }
 }
-
-
-
-
